DXE-544 Rename package `ivm` to `imaging`
