+++ pkg/imaging/policy_test.go
@@ -1,4 +1,4 @@
-package ivm
+package imaging
 
 import (
 	"context"
